A tool that relocates the paths of a database. Use it carefully.
The paths of sources are relative in the database. If you have an old
database with some absolute paths, and/or you want to move the
database (the .papiers.db file), you'll can use this tool to properly
relocate the paths of the sources so they are relative to the good
location.